+++ crates/cli/tests/reference/import-target-deno.js
@@ -25,7 +25,7 @@ function __wbg_get_imports() {
             const ret = arg0.get();
             return ret;
         },
-        __wbg_my_function_d6c76e66f4e6751a: function() {
+        __wbg_my_function_9c9b44befa10c96f: function() {
             b.my_function();
         },
         __wbg_new_625e056221ff6a6e: function(arg0) {
@@ -35,14 +35,14 @@ function __wbg_get_imports() {
         __wbg_no_catch_be4131677910f8cc: function() {
             no_catch();
         },
-        __wbg_reload_cbb85f596c80d974: function() {
+        __wbg_reload_2653607e441f1ab1: function() {
             window.location.reload();
         },
         __wbg_static_accessor_CONST_85b96acb48be57e1: function() {
             const ret = a.CONST;
             return ret;
         },
-        __wbg_write_1411fb0c6d3ebc7b: function(arg0, arg1) {
+        __wbg_write_ddba260cf11052e4: function(arg0, arg1) {
             window.document.write(getStringFromWasm0(arg0, arg1));
         },
         __wbindgen_init_externref_table: function() {