+++ crates/cli/tests/reference/targets-target-deno.js
@@ -12,7 +12,7 @@ export function add_that_might_fail(a, b) {
 function __wbg_get_imports() {
     const import0 = {
         __proto__: null,
-        __wbg_random_fcf922927450bad6: function() {
+        __wbg_random_59023cc6873cac50: function() {
             const ret = Math.random();
             return ret;
         },