+++ crates/cli/tests/reference/targets-target-no-modules-mvp.js
@@ -17,7 +17,7 @@ let wasm_bindgen = (function(exports) {
     function __wbg_get_imports() {
         const import0 = {
             __proto__: null,
-            __wbg_random_c82d91f28994c195: function() {
+            __wbg_random_05fd70d9b60b78cc: function() {
                 const ret = Math.random();
                 return ret;
             },